PR FIX + move component to dedicated YPP directory
